feat: CategoryClassifier에 카테고리 변화가 생기면 sidebar가 on 되도록 변경 (#152)
